Updated the answers for the math questions
Updated the answers to the math questions by using the random variable (with numbers ranging from 1-100 and updated comments to make it easier for me to understand what I did to my code later on. 

Continue working on the button functionality to make sure that when I click a certain button, it tells me if I'm 'right' or 'wrong'. 


If I am 'right', it will take me to the next question. If I am 'wrong', I will have to return to the home screen and start again (like mentioned in the flow chart) - Make sure to try and make that function using 'mousePressed' and to define it!